License Change
Hello contributors!

As an open source community, we should be able to make our products commercially viable without compromising our values. At the moment Infrared is licensed under GLP-3.0 which makes it difficult for companies to actually use Infrared. Therefore I appeal to all contributors of Infrared to agree to this new license (Apache 2.0) in order not to limit it in its growth. The change of the license has otherwise no other influence on the project, but only relaxes condition for commercial use.
If you all agree to this, I ask each of you to agree to this pull request.

Thank you very much and have an amazing day,
Hendrik Schlehlein aka. Haveachin